Move `utils.py`.  Delete unused code.
